Fixed problem with processing of previous aload in rewriteMonitorexit().
RewriteManager now parses XML files to build the list of methods that indirectly access aggregated state.

Updated frame model to ignore when the names of local variables become unknown.  It was never important, but could cause problems at the end of method bodies.

Had to add jaxb libraries to various projects for XML parsing.  These are only needed for Java 5 environemnts.  The projects and plug-ins need to be updated to reflect this.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/sierra-pmd@12153 d02ae147-ee12-4a35-9452-91c89e8b2592